Restructuring GUI into subdirs
