+++ intake/gui/catalog/__init__.py
@@ -1,0 +1,3 @@
+from .select import CatSelector
+from .add import CatAdder
+from .search import Search
