+++ intake/gui/icons/logo.png
