+++ intake/gui/source/__init__.py
@@ -1,0 +1,3 @@
+from .select import SourceSelector
+from .description import Description
+from .defined_plots import DefinedPlots
